+++ CHANGELOG.md
@@ -19,6 +19,7 @@
  - HotMovies: Search and scraping of ratings was fixed
  - Fixed crash if custom movie scraper was used with adult scrapers
  - Kodi NFO files will only contain `<uniqueid>` if there are valid ones
+ - TMDb TV now scrapes the full cast of a TV show instead of only the last season's (#1454)
 
 ### Changes